minor fixes

+++ converter_functions/write_ouxx.c
@@ -6,7 +6,7 @@
 /*   By: rsanchez </var/mail/rsanchez>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/23 04:53:37 by rsanchez          #+#    #+#             */
-/*   Updated: 2020/11/23 04:53:39 by rsanchez         ###   ########.fr       */
+/*   Updated: 2020/11/23 22:04:40 by romain           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,23 +14,23 @@
 
 void	write_o(va_list *param, t_flags *flags)
 {
-	unsigned_noboun(flags, va_arg(*param, unsigned long),
+	unsigned_noboun(flags, va_arg(*param, unsigned int),
 						"01234567", 8);
 }
 
 void	write_u(va_list *param, t_flags *flags)
 {
-	unsigned_noboun(flags, va_arg(*param, long), "0123456789", 10);
+	unsigned_noboun(flags, va_arg(*param, unsigned int), "0123456789", 10);
 }
 
 void	write_x(va_list *param, t_flags *flags)
 {
-	unsigned_noboun(flags, va_arg(*param, unsigned long),
+	unsigned_noboun(flags, va_arg(*param, unsigned int),
 						"0123456789abcdef", 16);
 }
 
 void	write_xx(va_list *param, t_flags *flags)
 {
-	unsigned_noboun(flags, va_arg(*param, unsigned long),
+	unsigned_noboun(flags, va_arg(*param, unsigned int),
 						"0123456789ABCDEF", 16);
 }
